==> notebooks/0.0-cw-review-data-in-brief.ipynb <==
# Lazada Product Review Data In Brief
----
## Description
First step of data exploration. Understanding data dan explore macro-characteristic of the data that we have to deal with

## Author
+ [Christian Wibisono](https://www.kaggle.com/christianwbsn)

import pandas as pd
import numpy as np

raw = pd.read_csv( "../data/raw/lazada_reviews.csv")

raw.head()

As we can see the data contain 2 columns *rating* and *review*

raw.info()

1. rating: numerical (1-5)
2. review: text
----
We have 280803 data points but there are 60570 missing values in review column.
Since our objective is to do **sentiment analysis** on text data, those missing value will be dropped later.
Those missing values couldn't tell us anything about the data

## Characteristic Of Marketplace Product Review in Bahasa Indonesia 

raw['review'][:10]

## 1.Informal words

print(raw['review'][2])
print("---------------")
print(raw['review'][7])

Data contain words in slang words such as:
+ *ga* --> tidak
+ *mantab* --> mantap
+ *nyampe* --> sampai

## 2. Abbreviation

print(raw['review'][10])
print("---------------")
print(raw['review'][11])

Data contain abbreviation i.e:
+ *knp* --> kenapa
+ *yg* --> yang
+ *tp* --> tapi

## 3. Foreign Word

print(raw['review'][30])
print("---------------")
print(raw['review'][19])

Data contain foreign languange i.e:
+ *seller* --> penjual
+ *packing* --> pengemasan
+ *recommended* --> direkomendasikan
+ *happy birthday* --> selamat ulang tahun

## 4. Domain Spesific Word

print(raw['review'][10])
print("---------------")
print(raw['review'][14])

Our domain: Marketplace Product Review 
Data contain domain spesific word i.e:
+ *flash sale*
+ *xiaomi*
+ *lazada*
+ *redmi*
+ *powerbank*

## 5. Emoji

print(raw['review'][12])
print("---------------")
print(raw['review'][31])

# 😃

This is the example of emoji found in our dataset

## 6. Emoticon

print(raw['review'][82])
print("---------------")
print(raw['review'][280323])

## :( sad or happy :)

## 7. Interjection

print(raw['review'][12983])
print("---------------")
print(raw['review'][222677])

Data contain interjection such as:
+ **wkwkwk**
+ **Hahahaha**

## 8. Typographical Error

print(raw['review'][15])
print("---------------")
print(raw['review'][8])

There are so **many** typographical error in our dataset. Single review can contain more than one typos
+ kadang**h** --> kadang(addition)
+ can**s**el --> cancel(replace)
+ pen**it**iman --> pengiriman(deletion,transposition,replace)
+ melalui**n** --> melalui(addition)

## 9. Irrelevant Text

print(raw['review'][1639])
print("---------------")
print(raw['review'][224544])
print("---------------")
print(raw['review'][80])

## 10. Completely Not in Bahasa

print(raw['review'][75])
print("---------------")
print(raw['review'][551])

## 11. Creative Isn't it?

print(raw['review'][38])
print("---------------")
print(raw['review'][21677])

## 12. Local *'Slang'* Words

print(raw['review'][17719])
print("---------------")
print(raw['review'][17025])

## 13. Unstandardized Words

print(raw['review'][0])
print("---------------")
print(raw['review'][5])

+ Sangatttttt --> sangat
+ lamaaaaaa --> lama
+ leleeeeeetttt --> lelet
+ ituuuuuuuu --> itu

## 14. Meaningless Words

print(raw['review'][224432])
print("---------------")
print(raw['review'][279396])
print("---------------")
print(raw['review'][279404])
print("---------------")
print(raw['review'][279403])

Know the meaning of those word? email me: [christian.wibisono7@gmail.com](mailto:christian.wibisono7@gmail.com)

## 15. Inconsistent Rating and Review

print("Rating " + str(raw['rating'][6970]) + ":" + raw['review'][6970])
print("---------------")
print("Rating " + str(raw['rating'][15908]) + ":" + raw['review'][15908])

Both condition appears in our dataset:
+ Positive review with negative rating
+ Negative review with positive rating

This condition can led to incorrectly labelled data

## 16. Reversed Words

print(raw['review'][228587])
print("---------------")

+ Kuy --> Yuk (reversed letter by letter)
