==> notebooks/1.3-cw-word-stemming.ipynb <==
import pandas as pd
import numpy as np
import string
import re
from collections import Counter

raw = pd.read_csv( "../data/interim/1.2_lazada_review_cleaner_pure.csv")

print(raw.info())

raw.head()

raw['review'] = raw['review'].apply(lambda x: x.replace(","," "))

from Sastrawi.Stemmer.StemmerFactory import StemmerFactory

factory = StemmerFactory()
stemmer = factory.create_stemmer()
raw['review'] = raw['review'].apply(lambda x: stemmer.stem(x).replace(" ",","))

def avg_word(words):
  return (sum(len(word) for word in words)/len(words))

raw['avg_word'] = raw['review'].apply(lambda x: avg_word(x.split(",")))

raw.to_csv("../data/interim/1.3_lazada_review_stemmed.csv", index=False)
