==> notebooks/3.0-cw-data-splitting-and-finalizing-processed-data.ipynb <==
import pandas as pd
import numpy as np

raw = pd.read_csv( "../data/interim/1.2_lazada_review_cleaner_pure.csv")
raw['review'] = raw['review'].apply(lambda x: x.replace(","," "))
sentences = raw.loc[:,['review','rating']]
def rating_to_sentiment(rating):
    if rating >= 3:
        return 1
    else:
        return 0
    
sentences['sentiment'] = sentences['rating'].apply(rating_to_sentiment)
sentences.drop(['rating'],axis=1,inplace=True)
neg = sentences[sentences['sentiment'] == 0]
pos = sentences[sentences['sentiment'] == 1]
print("All Processed Data")
print("Negative: ",neg.shape)
print("Positive: ",pos.shape)

from sklearn.model_selection import train_test_split
train, validation = train_test_split(sentences, test_size=0.2, random_state=42)
test = validation.loc[:,['review']]

train.head()

print(train.shape)

train.to_csv("../data/processed/train_1.csv", index=False, quoting=3)

test.head()

print(test.shape)

test.to_csv("../data/processed/test_1.csv", index=False, quoting=3)

validation.to_csv("../data/processed/validation_1.csv", index=False, quoting=3)

raw = pd.read_csv( "../data/interim/1.2_lazada_review_cleaner_stop_removed.csv")
raw['review'] = raw['review'].apply(lambda x: x.replace(","," "))
sentences = raw.loc[:,['review','rating']]
    
sentences['sentiment'] = sentences['rating'].apply(rating_to_sentiment)
sentences.drop(['rating'],axis=1,inplace=True)
neg = sentences[sentences['sentiment'] == 0]
pos = sentences[sentences['sentiment'] == 1]
print("All Processed Data")
print("Negative: ",neg.shape)
print("Positive: ",pos.shape)

train, validation = train_test_split(sentences, test_size=0.3, random_state=42)
test = validation.loc[:,['review']]

train.head()

print(train.shape)

train.to_csv("../data/processed/train_2.csv", index=False, quoting=3)

test.head()

print(test.shape)

test.to_csv("../data/processed/test_2.csv", index=False, quoting=3)

validation.to_csv("../data/processed/validation_2.csv", index=False, quoting=3)

raw = pd.read_csv( "../data/interim/1.3_lazada_review_stemmed.csv")
raw['review'] = raw['review'].apply(lambda x: x.replace(","," "))
sentences = raw.loc[:,['review','rating']]
    
sentences['sentiment'] = sentences['rating'].apply(rating_to_sentiment)
sentences.drop(['rating'],axis=1,inplace=True)
neg = sentences[sentences['sentiment'] == 0]
pos = sentences[sentences['sentiment'] == 1]
print("All Processed Data")
print("Negative: ",neg.shape)
print("Positive: ",pos.shape)

train, validation = train_test_split(sentences, test_size=0.3, random_state=42)
test = validation.loc[:,['review']]

train.head()

print(train.shape)

train.to_csv("../data/processed/train_3.csv", index=False, quoting=3)

test.head()

print(test.shape)

test.to_csv("../data/processed/test_3.csv", index=False, quoting=3)

validation.to_csv("../data/processed/validation_3.csv", index=False, quoting=3)
