==> notebooks/4.3-cw-fasttext.ipynb <==
import pandas as pd
import numpy as np

train = pd.read_csv( "../data/processed/train_1.csv")
test = pd.read_csv("../data/processed/test_1.csv")
validation = pd.read_csv("../data/processed/validation_1.csv")

from sklearn.model_selection import train_test_split
X = train['review']
y = train['sentiment']
X_train, X_test, y_train, y_test = train_test_split(X,y,test_size=0.2, random_state=42)

from gensim.models import KeyedVectors
id_model = KeyedVectors.load_word2vec_format('../data/external/wiki.id.vec')
words = []
for word in id_model.vocab:
    words.append(word)

# Printing out number of tokens available
print("Number of Tokens: {}".format(len(words)))

# Printing out the dimension of a word vector 
print("Dimension of a word vector: {}".format(
    len(id_model[words[0]])
))

# Print out the vector of a word 
print("Vector components of a word: {}".format(
    id_model[words[0]]
))
